chore: updated Tailwind to v4
