+++ assets/package.json
@@ -1,6 +1,7 @@
 {
   "devDependencies": {
-    "@tailwindcss/typography": "^0.5.10"
+    "@tailwindcss/forms": "^0.5.10",
+    "@tailwindcss/typography": "^0.5.16"
   },
   "dependencies": {
     "glider-js": "^1.7.9",
