+++ lib/lanttern_web/components/layouts/root.html.heex
@@ -7,13 +7,6 @@
     <.live_title suffix=" • Lanttern">
       <%= assigns[:page_title] || gettext("Hey!") %>
     </.live_title>
-    <!-- Google Fonts -->
-    <link rel="preconnect" href="https://fonts.googleapis.com" />
-    <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin />
-    <link
-      href="https://fonts.googleapis.com/css2?family=Montserrat:ital,wght@0,100..900;1,100..900&family=Open+Sans:ital,wght@0,300..800;1,300..800&family=Source+Code+Pro:ital,wght@0,200..900;1,200..900&display=swap"
-      rel="stylesheet"
-    />
 
     <link phx-track-static rel="stylesheet" href={~p"/vendor/glider.min.css"} />
     <link phx-track-static rel="stylesheet" href={~p"/assets/app.css"} />
